implement book an appointment with selected date and time

+++ admin/delete-user-account.php
@@ -5,18 +5,34 @@
             location.href='login.php'
         </script>";
     }
-    $ID = $_GET['id'];
-
+    $ID = $_POST['delete'];
+    $password = $_POST['adminPassword'];
+    $adminPassword = "";
+    
     include '../db_connection.php';
     mysqli_select_db($con, 'User');
-    
-    $sqlDeleteAccount = "DELETE FROM accounts WHERE id=$ID";
-    $sqlDeleteAccountInfo = "DELETE FROM user_info WHERE id=$ID";
+    $getAdminInfo = "SELECT * FROM user.admin WHERE ID = 1";
+    $admin = mysqli_query($con, $getAdminInfo);
+    if ($row = mysqli_fetch_array($admin)){        
+        $adminPassword = $row['Password'];
+    }
 
-    mysqli_query($con, $sqlDeleteAccount);
-    mysqli_query($con, $sqlDeleteAccountInfo);
-    echo "<script>
-            location.href='success.php';
-        </script>";
+    if ($password === $adminPassword){
+        $sqlDeleteAccount = "DELETE FROM accounts WHERE id=$ID";
+        $sqlDeleteAccountInfo = "DELETE FROM user_info WHERE id=$ID";
+
+        mysqli_query($con, $sqlDeleteAccount);
+        mysqli_query($con, $sqlDeleteAccountInfo);
+        echo "<script>
+                location.href='success.php';
+            </script>";        
+    } else {
+        echo "<script>
+            alert('incorrect Password');
+            location.href='users-account.php';
+            
+        </script>";    
+    }
+
     mysqli_close($con);
 ?>
